+++ bundle-polkadot-hw-ledger.js
@@ -9408,7 +9408,7 @@
 		hasRequiredPackageInfo = 1;
 		Object.defineProperty(packageInfo, "__esModule", { value: true });
 		packageInfo.packageInfo = void 0;
-		packageInfo.packageInfo = { name: '@polkadot/hw-ledger-transports', path: typeof __dirname === 'string' ? __dirname : 'auto', type: 'cjs', version: '13.5.1' };
+		packageInfo.packageInfo = { name: '@polkadot/hw-ledger-transports', path: typeof __dirname === 'string' ? __dirname : 'auto', type: 'cjs', version: '13.5.2' };
 		return packageInfo;
 	}
 
@@ -9477,6 +9477,7 @@
 	    encointer: 'Encointer',
 	    frequency: 'Frequency',
 	    integritee: 'Integritee',
+	    liberland: 'Liberland',
 	    polimec: 'Polimec',
 	    vara: 'Vara'
 	};