+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 13.5.2
 - api, api-contract, types 16.2.1
 - extension-dapp 0.59.2
 - phishing 0.25.12